[PLAT-6242][PLAT-6243] Implement gRPC equivalent of SSH RemoteShell to communicate with node agent + ansible over gRPC
Summary:
This adds the capability to run shell commands and ansible over node-agent. This is the starting point for platform for node-agent communication.
The feature is disabled by a feature flag to not disturb the existing flow until it is fully ready.

Ansible over gRPC is done as it can make the node-agent available faster to customers.

Test Plan:
Tested with the feature on.
1. Pause Universe with a combination of nodes with and without node-agent.
2. Start/Stop actions

Reviewers: amalyshev, sanketh, muthu, nbhatia

Subscribers: yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D20996